pyproject.toml
[tool.poetry]
name = "booklog-api"
version = "0.1.0"
description = "booklog for developer"
authors = ["이주환 <118493627+LeeJuHwan@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "booklog_api"}]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
pyyaml = "^6.0.1"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
dj-rest-auth = "^5.1.0"
django-autoslug = "^1.9.9"
django-taggit = "^5.0.1"
pillow = "^10.3.0"
django = "4.2.11"
pre-commit = "^3.7.0"


[tool.poetry.group.test.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py39']
force-exclude = 'migrations'

[tool.isort]
profile = "black"
filter_files = true
skip_glob = ["*/migrations/*", "*/config/*"]
